try tap encoder

+++ llava/model/multimodal_encoder/tokenize_anything/build_model.py
@@ -155,8 +155,9 @@ def image_tokenizer_for_llava(image_encoder, checkpoint=None, device=0, dtype="f
         #    vocab_size=text_tokenizer.n_words,
         #),
     )
-    model.float()
+    # model.float()
     load_weights_(model, checkpoint)
+    model = model.half()
     # model = model.to(device=get_device(device))
     # model = model.eval() if not kwargs.get("training", False) else model
     # model = model.half() if dtype == "float16" else model
